TQ: Add a liveness check to the cluster test (#9031)
This PR adds a liveness check at the end of each cluster proptest run.
It ensures the last generated configuration commits at all nodes, loads
the latest rack secret at all nodes and compares they are equal. If the
last generated configuration has aborted then we clone it and try again
with an epoch bump.

In order to complete the livness check, `PrepareAndCommit` support was
added. This is the last outstanding Node operation/message required for
the TQ protocol proper. The only remaining Node operations/messages
relate to LRTQ upgrade.

This builds on #8995