[sled-agent] Don't block `HardwareMonitor` when starting the switch zone if MGS isn't reachable (#9002)
See #8970 for context. This doesn't fix _all_ the ways we could get
wedged if the switch zone is unhealthy after starting it up, but it does
fix the case where it's so broken not even MGS is functional.

The changes here are pretty precarious (and took a bunch of retries to
get something that worked!). I ran them on dublin while doing some
testing for #8480, and was successfully able to start and stop the
switch zone even if the sidecar was powered off and MGS was
unresponsive.