sled-agent: begin incorporating Cosmo support (#9016)
Extend the existing `Baseboard` type to incorporate Cosmo, and introduce
a new `OxideSled` enum to represent different Oxide sled types so that
a) we can differentiate between different kinds of sled, and b)
parameterize omicron code with constant data specific to a sled type.

Also, identify places where we call the `is_gimlet` method, and instead
call that, `is_oxide_sled` (none of these are actually specific to
Gimlets in particular).

Find places where we use symbolic constants named `Gimlet` that should
be `Sled`; if these bleed into config files, set up Serde aliases to
accommodate a change.

Make it is that, if we set `skip_timesync` in a config file, we
propagate that to RSS config as well.

With these changes, we have successfully used the `gimlet-standalone` to
boot the control plane on a Cosmo. (Yes, the config name is somewhat
ironic, and should be addressed at some point.)