`InstanceUpdate` body: change `Option` fields to `Nullable` so they're required to be present (#9046)
Optional fields on a request body can be left out completely from the
JSON. In the case of instance update (and probably many other endpoints)
this is a problem because passing `null` and passing nothing have the
same effect, namely unsetting that field if it is already set in the DB.
And because the fields are optional, the generated client types do not
enforce that they are present (at least when creating the JSON by hand
and also in TypeScript, which distinguishes between optional and
nullable at the type level). It is easy to do this by accident: while
working on [this](https://github.com/oxidecomputer/console/pull/2900), I
[discovered a
bug](https://github.com/oxidecomputer/console/pull/2900#discussion_r2357519012)
in the console where we accidentally unset the auto restart policy on an
instance when you resize an instance.

This PR changes `Option` to `Nullable`, which I added in #8214 to fix
this problem: `null` remains a valid value, but the client is no longer
allowed to leave that field out of the body. If we don't want to do
this, I can live with it because in the console, I have [made a
helper](https://github.com/oxidecomputer/console/pull/2900/commits/60ca75e6827ae810dba1d53f87a70cf248f4f942)
that enforces that all fields are explicitly present. But I do think the
status quo is very error-prone.


https://github.com/oxidecomputer/omicron/blob/c6989117c72ecf88d4e3dc8a597d9fe703152a93/common/src/api/external/mod.rs#L3564-L3570