Support different types of Disk
In the same way that a higher level type was added to SiloUser and
SiloGroup, add a type to Disks. Currently this is only Crucible, but
that will change imminently to include local storage.

The main driver for this change was for different types of disks to
occupy the same PCI BDF space. Having a higher level "Disk" that
includes the runtime state for a disk (with information like what
instance it's attached to, what slot it occupies) be separate from where
the disk's blocks are coming from achieves this without any complicated
database or diesel magic.

Call sites that used the Disk db model now take either a Disk datastore
type, or directly take a CrucibleDisk (for functions that should only be
performed on Crucible disks). Several sagas have changed to only accept
CrucibleDisk in the parameters as well.

In this commit there's also a small change to how the instance spec is
computed with respect to the storage that provides a new option for
instances to have file backed storage. This is currently unused but the
framework for creating different components depending on this new Disk
type is there.

Other stuff in this commit:

- fix a bug where `project_create_disk` would insert a disk into the
  database, and _then_ perform some checks. If this failed in the
  context of a saga it would leave behind the inserted disk.

- delete `disk_refetch`, fulfilling a `TODO-cleanup`. similarly,
  `disk_set_runtime` is deleted. both of these will need to be revisited
  for hotplug, but the whole redesign of how sled agent and nexus work
  with VMMs obviates the old code.