fix: avoid strict argument evaluation to .then_some
An instance of `.then_some` on a bool caused an error because
`.then_some` takes a 'T' as an argument, and is thus evaluated
strictly; an error in the resulting eager evaluation caused an
error return, regardless of whether the predicating boolean
was true or not.

Nominally, using `.then` would have fixed it, but one cannot
conveniently use `?` in a closure, which the code we wanted
to run did.

So switch to a good 'ol `if` expression instead.

In testing, all of our configurations had the file we were testing
the presence of before reading.  But other folks do not.

However, one cannot conveniently use `?` in a closure in this
context, so

Fixes #9051