[3/n] [reconfigurator-planning] add and use set_target_release_and_old_repo (#8938)
For update-related tests, we're going to move to a world where the
initial system we start with has a target release set on it. We want to
ensure that as part of these tests, we transition from the initial
target release to the new one set up by these tests.

Add a setter to the `SystemDescription` called
`set_target_release_and_old_repo`, and use that in all the
update-related tests. As a nice benefit, we don't have to guess that the
first non-initial target release generation is 2 (in fact it will soon
become 3 since we'll initialize the system with a target release).