fix: use lazy argument evaluation to handle errors (#9053)
An instance of `.then_some` on a bool should have been `.then`. The
former evaluates its arguments eagerly, so runs _regardless of the value
of the bool_, while the latter does so lazily, which was the desired
behavior.

In testing, all of our configurations had the file we were testing the
presence of before reading. But other folks do not.

Fixes #9051