Store `PlanningReport`s in database for debugging only (PR 1/2) (#9029)
them in the database all along, but have punted on doing so because:

1. They're highly structured, making them difficult to map to SQL /
diesel
2. We expect to want to change them pretty frequently at least in the
near/mid term, which compounds problem 1

Chatting with @davepacheco, we came up with a compromise:

* These reports are not intended for programmatic consumption, and are
primarily useful as a debugging aid. (We want to report _something_ to
operators, but not at this level of detail. More on this momentarily.)
* Therefore, it seems okay to punt on the diesel / SQL problem by
storing these as serialized JSON blobs, under the condition that we
never attempt to parse them in Nexus proper, and therefore don't have to
deal with versioning.
* We can (and this PR does) add an `omdb` subcommand that _attempts_ to
parse them; if omdb and the Nexus that produced the report are out of
sync, this parsing may fail, but that's okay in that we can still at
least dump the raw JSON out. This PR hedges on this a bit by storing the
git commit in the JSON blob, so at least omdb can warn if it's expected
that it can't parse it (and note the git commit from which the support
operator would need an omdb to parse it successfully).