[4/n] [reconfigurator-planning] add a way to set the target release at system creation time (#8939)
As part of the work in #8726, updates will only happen if all zones are
at known versions before the update starts. In tests, the easiest way to
achieve this is to start off the `ExampleSystem` with a target release
at creation time.

This PR:

* adds support for that
* adds a simple `0.0.1` fake manifest that tests can use
* converts one of the update-related tests over to starting off with a
target release, as an example. (More tests will be converted over in
subsequent commits.)

In order to do this, the `ExampleSystemBuilder` now needs to be aware of
how to extract TUF repos. Reconfigurator-cli already had code to create
and extract TUF repos, which I've moved into
nexus-reconfigurator-planning.

This is going to run into issues with Cosmo boards since they'll have
multiple host phase 1 artifacts -- added TODOs to flag that.