Make wicketd + installinator cosmo aware (#9038)
This is the minimum set of changes to be able to update the next
generation sled Cosmo. The most intrusive change is choosing a separate
host phase 1 artifact.

This change does not attempt to differentiate between a Gimlet and a
Cosmo at the inventory level which will need to happen separately. This
change also does not support updating cosmos via reconfigurator driven
update.