Upload JUnit files from Buildomat to Trunk (#9049)
To help triage and handle flaky tests, it was suggested in the control
plane huddle a few weeks ago to setup [Trunk's flaky test
detection](https://trunk.io/flaky-tests). We run tests on Buildomat
though, which doesn't support setting the secret to authenticate with
Trunk.

To work around the issue, I created a new script to upload all the
relevant metadata to Trunk given the GitHub Check Run ID of a Buildomat
job:

```
tools/upload_buildomat_junit_to_trunk.sh <github-check-run-id>
```

The script is also hooked up to a GitHub Actions job that runs after
every completed Buildomat job. Note that if jobs change names or are
added the `case` block in the script will need to be updated.