PlanningReport: first pass at operator notes
These should be (a) serialized in the db along with the blueprint they
came from and (b) be available in the external API, but I wanted to get
a first pass out so folks could look over the way I've pulled these out.
I'm not at all sure this is the best way to generate these strings, and
even if it is, they undoubtedly need some serious wordsmithing for
external consumption and consistency.