[SCIM 1/3]: Add endpoint stubs
First, add endpoint stubs so that console and cli work can start. Use
the recently opened scim2-rs repo's structs to create all necessary
scim_v2_* endpoints, and also add endpoints for managing the SCIM client
bearer tokens (used to authenticate for those SCIM v2 endpoints).

The audit logging for the SCIM client bearer token CRUD will use the
opctx of the silo admin, but the SCIM v2 endpoints do not have an opctx
because they will never authenticate to an Actor: instead, create a new
"external-scim" user that will be used for the audit log, and will have
(in the 2/3 PR) the necessary permissions to authenticate SCIM clients.