[reconfigurator-planning] allow adding zones if target release generation is 1 (#9066)
For customers that are going to continue relying on MUPdate, the planner
should act the same way as it did before self-service update existed. We
ascertain this by looking at whether a target release has ever been set.

Most of the tests no longer require the
`add_zones_with_mupdate_override` config, so add a new
reconfigurator-cli script which specifically tests that config.