[nexus] decrease auto-restart cooldown to 5 minutes
When an instance is automatically restarted, it will not be permitted to
restart again until a cooldown period has elapsed since the previous
restart. Currently, that cooldown period is one hour, in order to
prevent situations where an instance that crashes every time it's
restarted is restarted in a hot loop. However, when an instance is on a
sled that is updated, it may then be restarted on another sled that is
also updated, and the cooldown timer will apply to the subsequent
restart. This means that during an update, some instances may not come
back for an hour, which is sad. Therefore, this commit decreases the
cooldown to 5 minutes, as discussed in #9094.

Fixes #9083