[reconfigurator] Planner should wait for all MGS updates before proceeding to zones (#9001)
Before any zones are updated, we need to make sure all MGS driven
updates have succeeded and/or are at the correct version. This PR makes
sure this is so, by bailing out if any MGS driven updates failed. All
skipped updates are also saved as part of the planner reports

Closes: https://github.com/oxidecomputer/omicron/issues/8285