Planner bg task: Save report even if we don't make a new target (#9088)
We talked about putting the background task's report in a watch channel,
and we might still want to do that if we decide we want to expose it in
the external API. But this change puts it into the _existing_ watch
channel that omdb can poke for debugging, and is a really trivial
change. With this it should be easy to compare the report of the current
target blueprint with the report of running the planner again after it's
realized.