misc: updates to run simulated omicron in container (#8996)
In https://github.com/oxidecomputer/terraform-provider-oxide/pull/497,
we're proposing to run the acceptance tests for the terraform provider
against simulated omicron. We're running omicron in a docker container
for portability. This PR contains a few small script changes to omicron
to make it work in a container:

* Optionally skip sudo, since some container environments don't include
it.
* Update the example config to listen on 0.0.0.0, so that we can reach
simulated omicron from outside the container.

---------

Co-authored-by: Josh Carp <joshcarp@mac.lan>