[gateway-client] replace SpType with version from gateway-types (#9093)
We'd like to derive `Arbitrary` on `SpType`, and the easiest way to do
so is to `replace` it.

One annoyance is that progenitor doesn't know that the replaced `SpType`
is `Copy`, so it changes generated type signatures to be `&SpType`
rather than `SpType`. A bit tedious to fix but not too bad overall.