[omdb] de-flake and tidy up `sp_ereport_ingester` status (#9054)
There's currently a test flake in the `test_omdb_success_cases` test due
to the output from the `sp_ereport_ingester` background task (see
#8979). This is because [the status output for that test will depend on
the number of MGSes and simulated SPs that are currently present when
the test runs][1]. In particular, this means that the output may [differ
based on whether the test is run in isolation or as part of a run
including other tests][2], as it's likelier that Nexus will have
discovered the simulated SPs when other tests have already run.

This branch fixes the flake by adding additional redaction configs for
this task's status output. We now redact all the numbers in the output.
Additionally, we must redact the sections listing individual SPs and
their ereport counts, and the section listing top level errors. Since
these sections are only shown when the lists of SPs or errors are
non-empty (the output looks weird if we print an empty section heading
for them), we don't use the existing `Redactor::section` method. This
method will replace the _indented contents_ of a section with the string
`<{section_name}_REDACTED>`. That's sufficient when just the section
_contents_ are variable, but in this case, the sections may or may not
be present depending on whether MGS was up, so we would get either a
redacted section (if it was present) _or_ nothing. Therefore, I've added
a new `Redactor::totally_annihilate_section` method that removes a
section _along with its heading_ for use in this case. It's a bit of a
shame to totally wipe these sections from the output, but it seemed like
the best choice under the circumstances.

I've also tweaked the status output a little bit to fix some cases where
numbers in the output were not aligned the way I wanted them to be, and
to make it slightly friendlier to the redactor by removing totals from
the headings of redacted sections.

Fixes #8979

[1]:
https://github.com/oxidecomputer/omicron/issues/8926#issuecomment-3229946163
[2]:
https://github.com/oxidecomputer/omicron/issues/8926#issuecomment-3235291262