TQ: Use BaseboardId instead of PlatformId
PlatformId was used as an internal placeholder during trust quorum
protocol development. In order to integrate the trust quorum protocol
code into the rest of Omicron we need to use the proper `BaseboardId`
type from `sled-agent-types`. The real `PlatformId` comes from the
`dice-util-msgs` crate and we'll need to be able to compare that
`PlatfromId` with `BaseboardId` and convert from `PlatformId` to
`BaseboardId` in the future.