Planner tests conversion CLI tests, for nexus generation testing (#9108)
In response to
https://github.com/oxidecomputer/omicron/pull/8936#discussion_r2369155140,
this PR migrates planner tests for nexus generation testing to use the
reconfigurator-cli.

To make this possible, the reconfigurator-cli has been modified to allow
the caller to modify the set of `active` and `not_yet` Nexus zones
explicitly. In production, these values would be read from the database,
and populated by the executor. Before this PR, these values were
inferred by the reconfigurator simulator, based on the target blueprint.