Nexus: Add an `inventory_loader` background task (PR 1/2) (#9148)
This adds a background task that periodically attempts to read the
latest inventory collection from the DB. It first does a query for just
the latest collection ID, then only does the full set of queries to load
the entire collection if the ID has changed from the last one it read.
Therefore, I set the period pretty aggressively (15 seconds), which is
what #5296 suggested. (Neither this PR nor the followup do the _rest_ of
#5296; i.e., changing how the `inventory_collector` task works.)

As of this PR, we only introduce the task and use it in tests. The
followup PR changes all the other background tasks that currently read
collections from the DB to read from this task's watch channel instead.