Nexus BG tasks: Use loaded inventory instead of collected inventory (PR 2/2) (#9149)
(This is the second half of #9148.)

As of this PR, all (I believe!) the background tasks that want to act on
the latest inventory collection now act on the latest inventory
collection loaded by the `inventory_loader` bg task. That task runs
frequently.

Fixes #8882. It's still possible two Nexuses could duel if one is acting
on a newer inventory collection, but that's always true. The window for
that should now be ~15 seconds, instead of waiting until the next time
_our own Nexus_ collects inventory.

I did not change other places in Nexus where we're reading inventory
outside of background tasks. I think it would probably make sense to
audit for those and make them read from this watch channel too? If that
sounds right I can file an issue.