[nexus] Update status API (#8897)
Closes #8869

* Add `update_status` endpoint at `/v1/system/update/status` that
includes:
  * Current target release (same as existing `target_release_view`)
* `time_last_progress`, representing the `time_made_target` of the
latest `bp_target` — meant to indicate the last time the update system
Did Something
* `components_by_release_version`, a map where the keys are to_string'd
`TufRepoVersion`s and the values are counts of components on that
version
* Remove `target_release_view` endpoint, which is fully redundant with
update status
* Add a line to the doc comment on `target_release_update` that you can
use update status to check the current target release
* Rework the `TargetRelease` struct that was previously being returned
from `target_release_view` and is now part of the update status response
* Remove `generation` from `TargetRelease` because it is not meaningful
to the end user
  * Other changes that are easier to explain in the inline comments

Listing blockers or problems preventing further update progress is
mentioned as a goal of #8869, but they are currently stored as JSON
blobs in the DB for debug purposes. They will soon be stored in a more
regimented way that should make it easy to stick a list of blockers in
the response body. If this PR gets merged without that, I'll make a
dedicated issue for it.

### Example status response

```json
{
  "target_release": {
    "time_requested": "2025-09-24T22:37:43.266338Z",
    "version": "2.0.0"
  },
  "components_by_release_version": {
    "install dataset": 7,
    "unknown": 15
  },
  "time_last_step_planned": "2025-09-24T22:37:41.556717Z",
  "suspended": false
}
```