Add `omdb nexus fetch-omdb` (#9172)
This PR adds a copy of the `omdb` binary into the Nexus zone. This
allows an `omdb` running in the switch zone to fetch a
potentially-different `omdb` binary from a running Nexus instance. In a
steady state this is useless, but during an update, the switch zone is
updated much earlier than the database schema or Nexus, leading to
difficulty observing the progress of the update. As a pretty cheap
workaround, shipping omdb with Nexus means we can always grab a version
that matches the currently-active Nexus (which will always match the
current database schema).

Closes #9075, which has more details about the kinds of errors we can
see mid-update.