[nexus] Add TUF repo list endpoint, make other endpoints conventional (#9106)
Built on top of https://github.com/oxidecomputer/omicron/pull/8897

These are breaking API changes and I won't merge until I have a PR up on
the CLI side that makes this all work.

### Add list endpoint and fix other ones

Made the endpoints match the other ones. Not married to the verb
`upload`, but `system_update_repository_update` felt a little silly.

```diff
-system_update_get_repository        GET      /v1/system/update/repository/{system_version}
-system_update_put_repository        PUT      /v1/system/update/repository
+system_update_repository_list       GET      /v1/system/update/repositories
+system_update_repository_upload     PUT      /v1/system/update/repositories
+system_update_repository_view       GET      /v1/system/update/repositories/{system_version}
```

### Standardize and simplify response types

What's responsible for this PR being so big is that I reworked these
endpoints to return a new `TufRepo` struct that is only the metadata and
no artifacts. The main reason for this was that the list endpoint would
have to pull artifacts for every repo in the list. The list is likely to
be small, so it's probably not a big deal, but it also seems that the
artifacts are not useful to the end user. Once the list endpoint was
simplified, it makes sense to return the same thing from the view and
update endpoints. The upload endpoint returns a `TufRepoUpload` which
has the `TufRepo` together with an indicator saying whether the repo
contents were new repo or already existed.