[nexus] re-enable `sp_ereport_ingester` post-R16 (#8838)
For R16, the `sp_ereport_ingester` background task was disabled in the
production Nexus config file (see #8709). This is because the
corresponding Hubris code for evacuating ereports from the SP had not
yet merged, resulting in Nexus yelling constantly about trying to
collect ereports from SPs that weren't listening on the ereport port.
Now, however, oxidecomputer/hubris#2126 has merged, and R16 has been
cut, so we can turn this back on. This commit does that.