Database support for delegating IP Pools for Oxide internal use (#9103)
- Add database queries for marking any number of IP Pools as delegated
for Oxide internal use, rather than assuming there's just one.
- Rework queries for linking / unlinking IP Pool and Silo, making them
concurrency-safe and preventing deletion when there are IPs or we're
trying to delete the last internal IP Pool.
- The public API does not change here. We're still assuming exactly one
IP Pool per IP version for internal usage, and not allowing the link or
pool to appear in the API. That'll be a follow-up commit.
- Fixes #8948