deflake some Crucible replacement tests (#9217)
This should fix flake opportunities in both:

*
`crucible_replacements::test_delete_volume_region_replacement_state_replacement_done`
*
`crucible_replacements::test_Delete_volume_region_replacement_state_requested`

This is my proposed change in #9216. From the logs of a flake and
careful reading it looks like we were almost-but-not-quite handling the
condition described in `wait_for_request_state` if the start and end
states for a replacement request are the same.

Consulting `get_region_replacement_by_id` and the returned replacement
request state tells us when the step has completed, in addition to
checking that the step had started.