Capture log file mtimes in support bundles
Currently the mtime for log files captured in support bundles is always
set as the default value of 1980-01-01T00:00. To more easily sort files
by time, it would be convenient to capture their actual mtimes in the
bundle zip.

Attempt to get the mtime for logs captured by `sled-diagnostics`,
falling back on the default if this fails, and copy the mtime into the
final bundle zip in nexus.