Add mapping of cubby to sled ID to support bundles (#9187)
Currently support bundles contain the sled UUID in file paths, and
serial number in `sled.txt`. However, when accessing the sled via the
tech port we generally refer to a sled by its cubby.

To make identification of sleds simpler, add a new `sled_info.json` file
to the bundle with a JSON-encoded mapping of sled serial to cubby and
UUID.