support-bundle storage: address lingering TODOs around transition from StorageManger to config reconciler (#9224)
This PR makes two changes; the first is small and does affect prod (but
should be very obviously correct), and the second is large but only
affects tests:

1. Replace `LocalStorage::dyn_datasets_config_list() ->
Result<DatasetsConfig, _>` with `LocalStorage::dyn_dataset_config()->
Result<DatasetConfig>`. The former returned an entire `DatasetsConfig`,
which no longer exists inside `OmicronSledConfig`, so the prod
implementation of the method had to create one. But the sole caller
really only wanted that to look up a single dataset, so replace the
method with one that looks up a single dataset.
2. Convert the tests from using a `LocalStorage` backed by
`StorageManagerTestHarness` to being backed by `sim::Storage`. This has
a couple pros and one con: pros are that we remove a reference to
`StorageManager` (which is no longer used in prod, but this isn't
obvious - ideally we could get rid of all the test consumers and then
remove it to avoid confusing) and that the tests can now run on
non-illumos systems; con is that we're no longer exercising true ZFS
operations in these tests. But I think the point of the `LocalStorage`
trait is that we don't need to exercise ZFS itself, so this con seems
pretty mild to me?

Using `sim::Storage` required adding some tracking of a (fake) `mounted`
property to the "datasets" it keeps track of in memory. I think this is
okay? But the nested dataset / dataset split in the APIs is still
confusing to me so I could easily have screwed something up in here.