Extract `ZfsTestHarness` from `StorageManagerTestHarness` (#9228)
After #9224 (on which this PR is based), there were two remaining users
of `StorageManagerTestHarness`, both of which rely on having real ZFS
pools / datasets available:

* The `zone_bundle` tests exercise production code that wants to
interact with ZFS properties (quota in particular; maybe others)
* The `sled-diagnostics::logs` tests use ZFS snapshots

However, `StorageManager` and the code around it is not in a great spot:
it appears to be real, production code, but it's no longer actually used
by sled-agent - all of its functionality has been absorbed into the
config reconciler.

This PR extracts a `ZfsTestHarness` - this is essentially a subset of
`StorageManagerTestHarness`, and copy/pastes its creation of vdevs for
real ZFS tests, but has a much simpler API: there were several methods
of `StorageManagerTestHarness` that had no callers, and it handles disks
/ datasets / zpools itself (in a very simple way) instead of handing
that off to `StorageManager`. Both of the test modules above are updated
to use `ZfsTestHarness`, which I _think_ clears the way to removing
`StorageManager` and `StorageManagerTestHarness`.