[reconfigurator] Stop minimally validating policy (#9241)
Rather than partially validating the policy, let it be whatever it wants
to be.

This avoids giving the impression that we are doing policy checking
prior to planning - we aren't, and will try to enact whatever policy
comes into the planner. But this is a feature, not a bug!

This PR is a mutually exclusive alternative to
https://github.com/oxidecomputer/omicron/pull/9239