Remove `StorageManager` and related dead code (#9229)
After #9228 (on which this PR is based), there are no remaining callers
of several sled-storage types: `StorageManager`,
`StorageManagerTestHarness`, `Error`, `AllDisks`, `StorageResources`,
and various other supporting types. This PR removes all of them.

The diff looks like there's a new `nested_dataset` module, but that's
really just a rename of what was left in the `manager` module after the
above code removal. Viewing the diffs of the two commits separately
makes that more clear.