clickhouse: update replication config in the correct file.
As a follow-up to #9183, update the replicated_max_ratio_of_wrong_parts setting
in the relevant file. After merging the previous PR, we noticed that the change
there didn't propagate to the config file in use on the rack. This patch
applies that same update to the correct config file in smf/.